+++ client/src/app/(authed)/notification/_ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./notification";
+export * from "./notification-header";
