+++ client/src/app/(authed)/notification/page.tsx
@@ -1,0 +1,12 @@
+import { Notification, NotificationHeader } from "./_ui";
+
+const Page = () => {
+  return (
+    <>
+      <NotificationHeader user_id={"abc"} />
+      <Notification />
+    </>
+  );
+};
+
+export default Page;
